refactor: :recycle: moved document types api
