--- src/entities/documents/model/types/documentTypesPageResponse.type.ts
@@ -1,11 +1,0 @@
-import type { DocumentTypeResponse } from './documentTypeResponse.type';
-
-export interface DocumentTypesPageResponse {
-  content: DocumentTypeResponse[];
-  empty: boolean;
-  first: boolean;
-  last: boolean;
-  numberOfElements: number;
-  totalElements: number;
-  totalPages: number;
-}
